3.1.3 SDK updates
update affdexsdk dependencies to use dynamic versioning with value
of "3.+", which will cause the sample apps to pick up the highest-numbered
SDK release in the 3.x series.  This way we don't need to update the samples
when we do a point release.

update gradle plugin to 2.2.2

set all versionCode values to 1 (normally a Bad Thing if you're app is
published to the Play store, but these aren't, so simpler to just have it
this way)

set APK naming style to appname-version-build#